Fix logic configuration
